requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
fastapi==0.115.12
flatbuffers==24.3.25
fonttools==4.51.0
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.2
h5py==3.11.0
idna==3.7
joblib==1.4.0
keras==3.3.2
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.8
numpy==1.26.4
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
pillow==10.3.0
protobuf==4.25.3
pycryptodome==3.21.0
pydantic==2.11.1
pydantic_core==2.33.0
pygame==2.5.2
Pygments==2.17.2
PyJWT==2.10.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.1
requests==2.31.0
rich==13.7.1
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
setuptools==78.1.0
six==1.16.0
sniffio==1.3.1
starlette==0.46.1
ta-lib==0.6.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
termcolor==2.4.0
threadpoolctl==3.4.0
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.2
wheel==0.45.1
wrapt==1.16.0
